test: RunnerPost 상세 조회시 Supporter 지원 여부 테스트 구현
