+++ backend/baton/src/test/java/touch/baton/assure/runnerpost/RunnerPostAssuredCreateTest.java
@@ -45,7 +45,7 @@ class RunnerPostAssuredCreateTest extends AssuredTestConfig {
                 .러너_게시글_생성_성공을_검증한다()
                 .생성한_러너_게시글의_식별자값을_반환한다();
 
-        final RunnerPostResponse.Detail 리뷰가_시작되지_않은_에단의_러너_게시글_Detail_응답 = 러너_게시글_Detail_응답을_생성한다(러너_에단, 러너_게시글_생성_요청, NOT_STARTED, 에단의_러너_게시글_식별자값, 1, 0L);
+        final RunnerPostResponse.Detail 리뷰가_시작되지_않은_에단의_러너_게시글_Detail_응답 = 러너_게시글_Detail_응답을_생성한다(러너_에단, 러너_게시글_생성_요청, NOT_STARTED, 에단의_러너_게시글_식별자값, 1, 0L, false);
         RunnerPostAssuredSupport
                 .클라이언트_요청()
                 .토큰으로_로그인한다(에단_로그인_토큰)
@@ -92,7 +92,8 @@ class RunnerPostAssuredCreateTest extends AssuredTestConfig {
                                                              final ReviewStatus 리뷰_상태,
                                                              final Long 러너_게시글_식별자값,
                                                              final int 조회수,
-                                                             final long 서포터_지원자수
+                                                             final long 서포터_지원자수,
+                                                             final boolean 서포터_지원_여부
     ) {
         return 러너_게시글_Detail_응답(
                 러너_게시글_식별자값,
@@ -104,6 +105,7 @@ class RunnerPostAssuredCreateTest extends AssuredTestConfig {
                 서포터_지원자수,
                 리뷰_상태,
                 true,
+                서포터_지원_여부,
                 러너,
                 러너_게시글_생성_요청.tags()
         );
